% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotcolor.R
\name{dotplotcolor}
\alias{dotplotcolor}
\title{thi  step to create a dotlpot indicating dat distrbitbution with z}
\usage{
dotplotcolor(datainput)
}
\arguments{
\item{datainput}{one set of data must be in number to work with}
}
\value{
none because it will create a plot
}
\description{
thi  step to create a dotlpot indicating dat distrbitbution with z
}
\concept{dotplot}
