% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikeg2.R
\name{maxlikg2}
\alias{maxlikg2}
\title{max like take sahpe of theta1 adn theta 2 finding log of bin pois function}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{is mean}

\item{theta2}{sd}

\item{lfun}{user input}

\item{...}{other labled}
}
\description{
max like take sahpe of theta1 adn theta 2 finding log of bin pois function
}
