% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{my Central Limit Function using runif, matrix, apply,creatin hist}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{looping}

\item{x}{is sample size}

\item{fun}{mode of operation}

\item{alpha}{forming confidence interval}

\item{cx}{cex plot}

\item{...}{labaled stuff}
}
\description{
my Central Limit Function using runif, matrix, apply,creatin hist
}
