% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysamplefuntion.R
\name{myf}
\alias{myf}
\title{Below is a code create bar plot of randomple sampling}
\usage{
myf(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{data}

\item{iter}{iterations or looping varaiables}

\item{time}{delay time between function executions}
}
\value{
barplot
}
\description{
Below is a code create bar plot of randomple sampling
}
