% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysamplefuntion.R
\name{myfunction}
\alias{myfunction}
\alias{myf}
\title{Below is a code create bar plot of random sampling}
\usage{
myf(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{data}

\item{iter}{iterations or looping variables}

\item{time}{delay time between function executions}
}
\value{
barplot
}
\description{
Below is a code create bar plot of random sampling
}
\concept{sampling}
