% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltmean.R
\name{newfunctionfromabove}
\alias{newfunctionfromabove}
\title{my Central Limit Function using runif, matrix, apply,creatin hist}
\usage{
newfunctionfromabove(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{is sample size}

\item{iter}{looping}

\item{a}{point a}

\item{b}{point b}
}
\description{
my Central Limit Function using runif, matrix, apply,creatin hist
}
