% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nticket.R
\name{ntickets}
\alias{ntickets}
\title{this is calculating ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seat}

\item{gamma}{probaility a plain will be truly over booked}

\item{p}{probaility of show}
}
\description{
this is calculating ntickets
}
