% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1.R
\docType{package}
\name{package1}
\alias{VINH4753SPR25}
\alias{VINH4753SPR25-package}
\alias{package1}
\title{premade dotplot}
\description{
creating a dotplot with hgiven data. note that the data must in number to work with.
}
\author{
Vinh Huynh \email{lyonkhang01@gmail.com}
}
