% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{create cruve with less than P(<=a)}
\arguments{
\item{mu}{is mean}

\item{sigma}{sd}

\item{a}{user input}
}
\value{
graph and area
}
\description{
create cruve with less than P(<=a)
}
\concept{curve}
